upd: added changes in the Order service, product service, customer service

+++ services/order-service/src/main/java/com/ecommerce/orderMS/dto/OrderItemDTO.java
@@ -9,6 +9,6 @@ public class OrderItemDTO {
 
     private String productId;
     private String productName;
-    private BigDecimal priceAtPurchase;
+    private Double priceAtPurchase;
     private Integer quantity;
 }
